Source code tidying and components renamed
