+++ app/responses/response_store.py
@@ -1,4 +1,5 @@
 from flask import session
+from abc import ABCMeta, abstractmethod
 
 RESPONSES = "resp"
 
@@ -9,19 +10,21 @@ class ResponseStoreFactory(object):
         return FlaskResponseStore()
 
 
-class IResponseStore(object):
-
+class AbstractResponseStore(metaclass=ABCMeta):
+    @abstractmethod
     def store_response(self, key, value):
         raise NotImplementedError()
 
+    @abstractmethod
     def get_response(self, key):
         raise NotImplementedError()
 
+    @abstractmethod
     def get_responses(self):
         raise NotImplementedError()
 
 
-class FlaskResponseStore(IResponseStore):
+class FlaskResponseStore(AbstractResponseStore):
 
     def store_response(self, key, value):
         if RESPONSES not in session:
