+++ tests/app/validation/test_required.py
@@ -1,11 +1,11 @@
-from app.validation.required import Required
+from app.validation.required_check import RequiredCheck
 import unittest
 
 
 class RequiredTest(unittest.TestCase):
 
     def test_required_validator(self):
-        required = Required()
+        required = RequiredCheck()
 
         # validate None
         result = required.validate(None)
