+++ tests/app/validation/test_type_validation_factory.py
@@ -1,6 +1,6 @@
 import unittest
 from app.validation.type_validator_factory import TypeValidatorFactory, TypeValidatorFactoryException
-from app.validation.integer import Integer
+from app.validation.integer_type_check import IntegerTypeCheck
 
 
 class TypeValidatorFactoryTest(unittest.TestCase):
@@ -10,6 +10,6 @@ class TypeValidatorFactoryTest(unittest.TestCase):
         validators = TypeValidatorFactory.get_validators_by_type('integer')
 
         self.assertEquals(len(validators), 1)
-        self.assertTrue(isinstance(validators[0], Integer))
+        self.assertTrue(isinstance(validators[0], IntegerTypeCheck))
 
         self.assertRaises(TypeValidatorFactoryException, TypeValidatorFactory.get_validators_by_type, "unknown_type")
